Fix #2 (implement parallel repository transfers)
The built-in multiprocessing module allows parallel processes. Since
downloading repositories is inherently I/O-bound, it makes sense to
parallelise the downloads – no conflicts are foreseen, as the processes
only write files with UUID names to a central location.

Now a significant portion of the time is spent pushing the grand
ensemble of files back up.